Criação de um template de dashboard para o menu principal Fixes #14
