+++ package.json
@@ -21,7 +21,8 @@
     "react-chartjs-2": "^3.0.3",
     "react-dom": "17.0.2",
     "react-grid-layout": "^1.2.5",
-    "sass": "^1.32.13"
+    "sass": "^1.32.13",
+    "zabbix-rpc": "^1.2.3"
   },
   "devDependencies": {
     "@types/node": "^15.0.1",
