+++ src/components/charts/chart.module.scss
@@ -1,9 +1,23 @@
 .chart {
+    display: absolute;
     width: 100%;
     height: 100%;
 
-    padding: 10px;
-
     background: white;
     border-radius: 5px;
+
+    border: 1px solid red;
+
+    button {
+        display: relative;
+        background: transparent;
+        border: 0;
+        float: right;
+
+        margin: 5px;
+
+        &:hover {
+            filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.6));
+        }
+    }
 }
